# pos-common/src/main/java/edu/nju/sa2022/micropos/models/Order.java
package edu.nju.sa2022.micropos.models;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;
import org.springframework.data.annotation.Id;

import java.util.Map;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class Order {

    @Id
    private String id;

    private String userId;

    private Map<String, Integer> items;
    private Double total;

}